User/arepikov/firmware cleanup (#71)
* Moved included libaries to dedicated folder

* removed stepper

* Create readme.md

* Fixed symlinks

* Fixed cmake error
