+++ lunabot_embedded/src/pb_common.c
@@ -1,1 +1,1 @@
-../firmware/teensy_main/lib/lunabot_drivers/pb_common.c
+../firmware/teensy_main/lib/lunabot_includes/pb_common.c
