+++ lunabot_embedded/src/pb_common.h
@@ -1,1 +1,1 @@
-../firmware/teensy_main/lib/lunabot_drivers/pb_common.h
+../firmware/teensy_main/lib/lunabot_includes/pb_common.h
