+++ lunabot_embedded/src/pb_decode.c
@@ -1,1 +1,1 @@
-../firmware/teensy_main/lib/lunabot_drivers/pb_decode.c
+../firmware/teensy_main/lib/lunabot_includes/pb_decode.c
