+++ lunabot_embedded/src/pb_decode.h
@@ -1,1 +1,1 @@
-../firmware/teensy_main/lib/lunabot_drivers/pb_decode.h
+../firmware/teensy_main/lib/lunabot_includes/pb_decode.h
