+++ lunabot_embedded/src/pb_encode.c
@@ -1,1 +1,1 @@
-../firmware/teensy_main/lib/lunabot_drivers/pb_encode.c
+../firmware/teensy_main/lib/lunabot_includes/pb_encode.c
