+++ lunabot_embedded/src/pb_encode.h
@@ -1,1 +1,1 @@
-../firmware/teensy_main/lib/lunabot_drivers/pb_encode.h
+../firmware/teensy_main/lib/lunabot_includes/pb_encode.h
